requirements.txt
hds-821pr==0.0.1
paho-mqtt==1.3.1
pyserial==3.4
